Final re-architecturing
Apparently flag stores flags globally, so, bad news if you have built
in cli support in the library... :(
